Update patch set 3
Patch Set 3:

okay, thanks.

so where is the "type"? I see image-id but that is not necessarily the type. Or is it? Are we going to have type identities somewhere so folks can expand it? 

I think the name "images" is too broad. Catalog.yang also has "images". IMO we need to be more explicit, otherwise is too confusing.

I think we also need a "oper" container + list that stores the running images managed by a certain driver.

Patch-set: 3